Improvement: Optimize RunEnd array filter for sparse masks
